Use Sdr for RfM light translation
Use USD's Shader Definition Registry for RfM Light Translation, which
allows us to avoid manually keeping code in sync with any associated
schemas. This change will still support the legacy method for now as
much of the Sdr capabilities are fairly recent additions to Usd.
